clean up tests with web helpers
