+++ spec/spec_helper.rb
@@ -8,6 +8,7 @@ require 'capybara/rspec'
 require 'rspec'
 require 'simplecov'
 require 'simplecov-console'
+require 'features/web_helpers'
 
 # tell Capybara about our app class
 Capybara.app = RPS
